// src/components/forms/Form/interface.ts
export type InternalNamePath = (string | number)[];
export type NamePath = string | number | InternalNamePath;

// eslint-disable-next-line @typescript-eslint/no-explicit-any
export type StoreValue = any;
export type Store = Record<string, StoreValue>;

export interface Meta {
  touched: boolean;
  validating: boolean;
  errors: string[];
  warnings: string[];
  name: InternalNamePath;
  rules: RuleObject[];
}

export interface InternalFieldData extends Meta {
  value: StoreValue;
}

// Used by `setFields` config
export interface FieldData extends Partial<Omit<InternalFieldData, 'name'>> {
  name: NamePath;
}

export type FieldsErrors = Record<string, Meta['errors']>;
// ...
export type RuleObject = AggregationRule | ArrayRule;

export type Rule = RuleObject | RuleRender;
export interface ValidateErrorEntity<Values = unknown> {
  values: Values;
  errorFields: { name: InternalNamePath; errors: string[] }[];
  outOfDate: boolean;
}

export interface FieldEntity {
  onStoreChange: (
    store: Store,
    namePathList: InternalNamePath[] | null,
    info: ValuedNotifyInfo,
  ) => void;
  isFieldTouched: () => boolean;
  isFieldDirty: () => boolean;
  isFieldValidating: () => boolean;
  isListField: () => boolean;
  isList: () => boolean;
  isPreserve: () => boolean;
  validateRules: (options?: ValidateOptions) => Promise<RuleError[]>;
  getMeta: () => Meta;
  getNamePath: () => InternalNamePath;
  getErrors: () => string[];
  getWarnings: () => string[];
  props: {
    name?: NamePath;
    rules?: Rule[];
    dependencies?: NamePath[];
    initialValue?: unknown;
  };
}

export interface FieldError {
  name: InternalNamePath;
  errors: string[];
  warnings: string[];
}

export interface RuleError {
  errors: string[];
  rule: RuleObject;
}

export interface ValidateOptions {
  triggerName?: string;
  validateMessages?: ValidateMessages;
  // Recursive validate. It will validate all the name path that contains the provided one.
  // e.g. ['a'] will validate ['a'] , ['a', 'b'] and ['a', 1].
  recursive?: boolean;
}

export type InternalValidateFields<Values = unknown> = (
  nameList?: NamePath[],
  options?: ValidateOptions,
) => Promise<Values>;

export type ValidateFields<Values = unknown> = (
  nameList?: NamePath[],
) => Promise<Values>;
// ...
export interface Callbacks<Values = any> {
  onValuesChange?: (changedValues: unknown, values: Values) => void;
  onFieldsChange?: (changedFields: FieldData[], allFields: FieldData[]) => void;
  onFinish?: (values: Values) => Promise<void> | void;
  onFinishFailed?: (errorInfo: ValidateErrorEntity<Values>) => void;
}

export type WatchCallBack = (
  values: Store,
  namePathList: InternalNamePath[],
) => void;

export interface InternalHooks {
  dispatch: (action: ReducerAction) => void;
  initEntityValue: (entity: FieldEntity) => void;
  registerField: (entity: FieldEntity) => () => void;
  useSubscribe: (subscribable: boolean) => void;
  setInitialValues: (values: Store, init: boolean) => void;
  destroyForm: () => void;
  setCallbacks: (callbacks: Callbacks) => void;
  registerWatch: (callback: WatchCallBack) => () => void;
  getFields: (namePathList?: InternalNamePath[]) => FieldData[];
  setValidateMessages: (validateMessages: ValidateMessages) => void;
  setPreserve: (preserve?: boolean) => void;
  getInitialValue: (namePath: InternalNamePath) => StoreValue;
  getInitialValues: () => Store;
}

export type SetErrorsOptions = {
  prefix?: string;
};

// Only return partial when type is not any
type RecursivePartial<T> = T extends object
  ? {
      [P in keyof T]?: T[P] extends (infer U)[]
        ? RecursivePartial<U>[]
        : T[P] extends object
        ? RecursivePartial<T[P]>
        : T[P];
    }
  : // eslint-disable-next-line @typescript-eslint/no-explicit-any
    any;

// eslint-disable-next-line @typescript-eslint/no-explicit-any
export interface FormInstance<Values = any> {
  // Origin Form API
  getFieldValue: (name: NamePath) => StoreValue;
  getFieldsValue: (() => Values) &
    ((
      nameList: NamePath[] | true | null,
      filterFunc?: (meta: Meta) => boolean,
      // eslint-disable-next-line @typescript-eslint/no-explicit-any
    ) => any);
  getFieldError: (name: NamePath) => string[];
  getFieldsError: (nameList?: NamePath[]) => FieldError[];
  getFieldWarning: (name: NamePath) => string[];
  isFieldsTouched: ((
    nameList?: NamePath[],
    allFieldsTouched?: boolean,
  ) => boolean) &
    ((allFieldsTouched?: boolean) => boolean);
  isFieldTouched: (name: NamePath) => boolean;
  isFieldValidating: (name: NamePath) => boolean;
  isFieldsValidating: (nameList: NamePath[]) => boolean;
  resetFields: (fields?: NamePath[]) => void;
  setError: (
    name: NamePath,
    error: Meta['errors'],
    options?: SetErrorsOptions,
  ) => void;
  setErrors: (error: FieldsErrors, options?: SetErrorsOptions) => void;
  setFields: (fields: FieldData[]) => void;
  setFieldValue: (name: NamePath, value: unknown) => void;
  setFieldsValue: (values: RecursivePartial<Values>) => void;
  validateFields: ValidateFields<Values>;

  // New API
  submit: () => void;
}

export type FieldType =
  | 'number'
  | 'checked'
  | 'file'
  | 'files'
  | 'select'
  | 'url'
  | 'multiSelect'
  | 'date';

export type InternalFormInstance = Omit<FormInstance, 'validateFields'> & {
  validateFields: InternalValidateFields;

  // Passed by field context props
  prefixName?: InternalNamePath;

  validateTrigger?: string | string[] | false;

  // Form component should register some content into store.
  // We pass the `HOOK_MARK` as key to avoid user call the function.
  getInternalHooks: (secret: string) => InternalHooks | null;

  /** @private Internal usage. Do not use it in your production */
  _init?: boolean;
};

// eslint-disable-next-line @typescript-eslint/no-explicit-any
export type EventArgs = any[];

type ValidateMessage = string | (() => string);
export interface ValidateMessages {
  default?: ValidateMessage;
  required?: ValidateMessage;
  enum?: ValidateMessage;
  whitespace?: ValidateMessage;
  date?: {
    format?: ValidateMessage;
    parse?: ValidateMessage;
    invalid?: ValidateMessage;
    max?: ValidateMessage;
    min?: ValidateMessage;
    range?: ValidateMessage;
  };
  types?: {
    string?: ValidateMessage;
    method?: ValidateMessage;
    array?: ValidateMessage;
    object?: ValidateMessage;
    number?: ValidateMessage;
    date?: ValidateMessage;
    boolean?: ValidateMessage;
    integer?: ValidateMessage;
    float?: ValidateMessage;
    regexp?: ValidateMessage;
    email?: ValidateMessage;
    url?: ValidateMessage;
    hex?: ValidateMessage;
  };
  string?: {
    len?: ValidateMessage;
    min?: ValidateMessage;
    max?: ValidateMessage;
    range?: ValidateMessage;
  };
  number?: {
    len?: ValidateMessage;
    min?: ValidateMessage;
    max?: ValidateMessage;
    range?: ValidateMessage;
  };
  array?: {
    len?: ValidateMessage;
    min?: ValidateMessage;
    max?: ValidateMessage;
    range?: ValidateMessage;
  };
  pattern?: {
    mismatch?: ValidateMessage;
  };
}